pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "line1 项目"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django==5.2.7",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.2.0",
    "ipython>=9.6.0",
    "ruff>=0.14.3",
]

[tool.ruff]
fix = true
ignore = ["Q000", "Q001", "Q002"]

[tool.ruff.format]
quote-style = "preserve"
